requirements.txt

aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.5
distro==1.9.0
dnspython==2.6.1
easyocr==1.7.1
et-xmlfile==1.1.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
googlesearch-python==1.2.4
greenlet==3.0.3
huggingface-hub==0.25.0
idna==3.7
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.7
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-text-splitters==0.0.1
langsmith==0.1.54
lazy_loader==0.4
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
meson==1.5.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pip==24.2
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.4
PyPDF2==2.10.5
pypdfium2==4.30.0
pyspark==3.5.2
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.9.11
reportlab==4.2.2
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==74.0.0
shapely==2.0.6
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
sympy==1.13.2
tabula-py==2.9.3
tabulate==0.9.0
tenacity==8.2.3
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
ultralytics==8.3.8
ultralytics-thop==2.0.9
urllib3==2.2.1
Werkzeug==3.0.2
XlsxWriter==3.2.0
yarl==1.9.4
